Rename package from conf to config and update related file paths
