+++ pkg/config/conf.go
@@ -1,7 +1,7 @@
-package conf
+package config
 
 import (
-	"neko-acm/utils"
+	utils "neko-acm/pkg/utils"
 )
 
 type Config struct {
