+++ pkg/config/var.go
@@ -1,4 +1,4 @@
-package conf
+package config
 
 var (
 	Conf *Config = &Config{}
